Migrate `Routes/TriggerMessage` story to TS
